Add max tokens options to Parser
Adds a `max_tokens` option to the parser which enforces a max tokens
limit. The lexer will stop once the max tokens limit has been exceeded
and return an error.

The parser has been changed to return a `ParseResult` object which
contains both the parsed AST and the `tokens_count`. This is exposed so
that applications can log the tokens count regardless if the limit is
exceeded or not and to determine a safe limit before enforcing it.