Return ParseDetails from Parse trait (#80)
The Parse trait now returns `ParseDetails` instead of a `Result` to more
accurately reflect that result information can always be returned and
not only for success cases.

This exposes the token count now regardless of success or failure. Even
when the parser returns an error, there's still a token count.